Add deterministic-CBOR infrastructure (GER-1965)
Pin swift-cbor at the same revision CoreAppLogic uses, and add a small
DeterministicCbor seam establishing the conventions new wire types must
follow: short string map keys (swift-cbor's Codable surface only ever
produces/accepts string keys — it silently drops non-string keys on
decode rather than erroring, so integer keys are wire-incompatible by
construction), the archive-companion pattern, and riding a CBOR blob
inside a positional container as a plain Data field.

Also hoists the hex-fixture decode helper out of AgentHelloDualOfferTests
into shared test support, for reuse by new wire-pinning tests.
