+++ Package.swift
@@ -25,6 +25,12 @@ let package = Package(
 			from: "0.2.2"
 		),
 		.package(url: "https://github.com/swift-libp2p/swift-bases.git", from: "0.2.0"),
+		.package(
+			// pinned by revision: `Options.deterministicCbor` (RFC 8949 §4.2.1)
+			// is on main and not in any released tag. Same pin as CoreAppLogic.
+			url: "https://github.com/nnabeyang/swift-cbor.git",
+			revision: "8d9b9c25284c6a2f0a564f4c42c7dc3466d08472"
+		),
 	],
 	targets: [
 		// Targets are the basic building blocks of a package, defining a module or a test suite.
@@ -36,6 +42,7 @@ let package = Package(
 				.product(name: "AtprotoTypesMocks", package: "AtprotoTypes"),
 				.product(name: "Base64", package: "swift-bases"),
 				"GermConvenience",
+				.product(name: "SwiftCbor", package: "swift-cbor"),
 			]
 		),
 		.target(
