+++ Tests/CommProtocolTests/IdentityExchange/AgentHelloDualOfferTests.swift
@@ -209,21 +209,3 @@ struct AgentHelloDualOfferTests {
 		+ "222222222222222222222222222222222222222222222222222222222222222222"
 		+ "22222222222241da1b5980000000"
 }
-
-extension Data {
-	fileprivate init?(hexString: String) {
-		guard hexString.count.isMultiple(of: 2) else { return nil }
-		var bytes = [UInt8]()
-		bytes.reserveCapacity(hexString.count / 2)
-		var index = hexString.startIndex
-		while index < hexString.endIndex {
-			let next = hexString.index(index, offsetBy: 2)
-			guard let byte = UInt8(hexString[index..<next], radix: 16) else {
-				return nil
-			}
-			bytes.append(byte)
-			index = next
-		}
-		self.init(bytes)
-	}
-}
